Rename startup input grid
