+++ mj-calc/src/modules/NameInputGrid.tsx
